Debounce. Reformat - codelite does something strange with tabs...

+++ NettigoKeypad.h
@@ -2,17 +2,18 @@
 //Published on MIT licence
 // (c) 2011 Nettigo
 #include "WProgram.h"
+
 #define NG_KEYPAD_SIZE  6
 
 class NG_Keypad
 {
   public:
-  static const byte NONE    = 0;
-  static const byte SELECT  = 1;
-  static const byte LEFT    = 2;
-  static const byte DOWN    = 3;
-  static const byte UP      = 4;
-  static const byte RIGHT   = 5;
+  static const int NONE    = 0;
+  static const int SELECT  = 1;
+  static const int LEFT    = 2;
+  static const int DOWN    = 3;
+  static const int UP      = 4;
+  static const int RIGHT   = 5;
   
   NG_Keypad(void);  
 	
@@ -24,26 +25,26 @@ class NG_Keypad
   // takes two args
   // key - which key should be used
   // userF - which function should be called when key pressed
-  int register_handler( byte key, void (*userF)(void) );
+  int register_handler( int key, void (*userF)(void) );
   
   //Check for handlers to be called. Takes as argument analogRead result.
-  void check_handlers(byte rd);
+  void check_handlers(int rd);
 	
-	//return current debounce delay
-	unsigned int getDebounceDelay( void );
+  //return current debounce delay
+  unsigned int getDebounceDelay( void );
 	
-	//set new debounce delay
-	void setDebounceDelay (unsigned int);
+  //set new debounce delay
+  void setDebounceDelay (unsigned int);
   
-	private:
+  private:
   void  (*_functions[NG_KEYPAD_SIZE])(void);
   
-	//default debounce timeout
-	unsigned int debounceDelay;
+  //default debounce timeout
+  unsigned int debounceDelay;
 	
-	//what last key was pressed and when (millis)
-	byte lastKey;
-	unsigned long lastKeyTime;
+  //what last key was pressed and when (millis)
+  int lastKey;
+  unsigned long lastKeyTime;
   
 };
